refactor(setup): Refactor file and directory names to adhere to naming standards v2
