+++ src/lib/components/admin/cardAnswers/default-CardAnswer.tsx
@@ -3,7 +3,7 @@ import Input from "../../common/input/defaultInput";
 import { Anwsers } from "@/lib/interface/question.interface";
 import { useDispatch } from "react-redux";
 import { addAnswer } from "@lib/state/answer/answerSlice";
-import DefaultCheckBoxAnswer from "./check-box-answer/defaultCheckBoxAnswer";
+import DefaultCheckBoxAnswer from "./checkBoxAnswer/defaultCheckBoxAnswer";
 
 interface DefaultCardAnswerProps {
   placeholder?: string;
