+++ src/lib/components/admin/createQuizer/createQuestion/defaultCreateQuestion.tsx
@@ -2,8 +2,8 @@ import React, { useEffect, useRef, useState } from "react";
 import Image from "next/image";
 import Form from "@lib/components/common/form/defaultForm";
 import Input from "@lib/components/common/input/defaultInput";
-import Select from "@lib/components/common/selects/defaultSelects";
-import ButtonDefault from "@lib/components/common/buttons/buttonDefaults";
+import Select from "@lib/components/common/selects/defaultSelect";
+import ButtonDefault from "@/lib/components/common/buttons/buttonDefault";
 import DefaultCardAnsswer from "../../cardAnswers/default-CardAnswer";
 import { Anwsers, Question } from "@/lib/interface/question.interface";
 import { useDispatch, useSelector } from "react-redux";
