+++ src/lib/components/admin/createQuizer/createQuizerByExcel/defaultCreateQuizerByExcel.tsx
@@ -1,4 +1,4 @@
-import ButtonDefault from "@lib/components/common/buttons/buttonDefaults";
+import ButtonDefault from "@/lib/components/common/buttons/buttonDefault";
 import { setTurnOffPopup, setTurnOnPopup } from "@lib/state/popup/popupSlice";
 import { addQuestions } from "@lib/state/questions/questionSlice";
 import { generateUniqueId } from "@/utils/generateUniqueId";
