+++ src/lib/components/admin/popupAdmin/defaultPopupAdmin.tsx
@@ -1,6 +1,6 @@
 import React, { useCallback, useEffect, useState } from "react";
 import Image from "next/image";
-import DefaultLoading from "@lib/components/common/loadings/defaultLoadings";
+import DefaultLoading from "@lib/components/common/loadings/defaultLoading";
 
 interface PropsPopupAdmin {
   popupDefault: boolean;
