+++ src/lib/components/common/sliders/defaulltSlider.tsx
