+++ src/lib/components/home-page/index.tsx
@@ -4,16 +4,16 @@ import Login from "./login/index";
 import { useRouter } from "next/router";
 import useLocalStorage from "@lib/hook/useLocalStorage";
 import Image from "next/image";
-import ButtonDefault from "../common/buttons/buttonDefaults";
+import ButtonDefault from "../common/buttons/buttonDefault";
 import { typeAccount } from "@lib/config/typeAccount";
 import Form from "../common/form/defaultForm";
-import Select from "@lib/components/common/selects/defaultSelects";
+import Select from "@lib/components/common/selects/defaultSelect";
 import Input from "../common/input/defaultInput";
 import countries from "@lib/config/countries";
 import { verifyLogin } from "@/api/auth.api";
 import { checkEmail, postUser } from "@/api/user.api";
 import { useForm } from "react-hook-form";
-import PopupDefault from "@lib/components/common/popups/popupWelcomes";
+import PopupDefault from "@lib/components/common/popups/popupWelcome";
 
 interface UserInfo {
   fullName: string;
