+++ src/lib/components/home-page/login/index.tsx
@@ -1,10 +1,10 @@
 import React, { useEffect, useState } from "react";
 import Image from "next/image";
 import { verifyLogin } from "@/api/auth.api";
-import PopupDefault from "@lib/components/common/popups/popupWelcomes";
+import PopupDefault from "@lib/components/common/popups/popupWelcome";
 import Form from "../../common/form/defaultForm";
 import Input from "../../common/input/defaultInput";
-import ButtonDefault from "../../common/buttons/buttonDefaults";
+import ButtonDefault from "../../common/buttons/buttonDefault";
 import { useRouter } from "next/router";
 import { useDispatch } from "react-redux";
 import { AppDispatch } from "@lib/state/store";
