+++ src/pages/admin/my-quizz/[id].tsx
@@ -1,10 +1,10 @@
-import ChartTwo from "@lib/components/common/chart/chartTwos";
+import ChartTwo from "@/lib/components/common/charts/chartTwo";
 import React, { useEffect, useState } from "react";
 import Image from "next/image";
 import { useRouter } from "next/router";
 import { getItemQuizzByUser, getListQuizz } from "@/api/quizz.api";
 import { getQuestionPercentagesByQuizzId } from "@/api/quizzResult.api";
-import ButtonDefault from "@lib/components/common/buttons/buttonDefaults";
+import ButtonDefault from "@/lib/components/common/buttons/buttonDefault";
 import Link from "next/link";
 import { debounce } from "@/utils/debounce";
 import { Anwsers } from "@/lib/interface/question.interface";
